+++ crates/swc_ecma_minifier/src/compress/optimize/props.rs
